rename Surface -> FrequentistSurface (and introduce base ISurface) in anticipation of a BayesianSurface cousin
